feat: initial extraction of header.rs
